+++ edgework/__init__.py
@@ -1,6 +1,6 @@
-"""Edgework NHL API Client - Version 0.4.6"""
+"""Edgework NHL API Client - Version 0.4.7"""
 
-__version__ = "0.4.5"
+__version__ = "0.4.7"
 
 from .edgework import Edgework